Some adjustments to make this work with the unknown_instances experiment
The main change here is enabling the unknown_instances experiment inside
the "ec2" module so that Terraform Core will permit the aws_instance
resource to have an unknown value as its for_each.

While testing this I found that (as has often been the case, even outside
of stacks) we can't rely on EC2 to return the subnets that exist if they
have only just recently been created, because the EC2 API is not
read-after-write consistent. Instead, we should pass the subnet ids from
one component to the other within Terraform, thereby ensuring that both
components agree on which subnets ought to exist.

The remaining parts not mentioned above are just small differences that
arose as I was re-genericing things so that this could be an example again,
removing the adaptations I made for my specific environment.